pgsql: Split layer.go to files in layer module
